# docs/openapi/api.v1.json
{
  "components" : {
    "parameters" : {
      "CorrelationId" : {
        "description" : "Optional trace ID; generated if absent and returned on every response",
        "in" : "header",
        "name" : "X-Correlation-Id",
        "required" : false,
        "schema" : {
          "example" : "demo-123",
          "type" : "string"
        }
      }
    },
    "schemas" : {
      "ApiError" : {
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "correlationId" : {
            "type" : "string"
          },
          "error" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "path" : {
            "type" : "string"
          },
          "status" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timestamp" : {
            "format" : "date-time",
            "type" : "string"
          },
          "violations" : {
            "items" : {
              "$ref" : "#/components/schemas/FieldViolation"
            },
            "type" : "array"
          }
        }
      },
      "DownloadLinkResponse" : {
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "url" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "FieldViolation" : {
        "properties" : {
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "PortalDocumentView" : {
        "properties" : {
          "category" : {
            "enum" : [ "DAMAGE_PHOTO", "REPAIR_ESTIMATE", "POLICE_REPORT", "INVOICE", "MEDICAL_REPORT", "OTHER" ],
            "type" : "string"
          },
          "claimId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "contentType" : {
            "type" : "string"
          },
          "fileName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "sizeBytes" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PENDING_UPLOAD", "UPLOADED", "PROCESSING", "PROCESSED", "FAILED", "REJECTED" ],
            "type" : "string"
          },
          "uploadedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "PortalUploadResponse" : {
        "properties" : {
          "document" : {
            "$ref" : "#/components/schemas/PortalDocumentView"
          },
          "upload" : {
            "$ref" : "#/components/schemas/UploadInstructions"
          }
        },
        "type" : "object"
      },
...
      "StaffCompletion" : {
        "properties" : {
          "document" : {
            "$ref" : "#/components/schemas/StaffDocumentView"
          },
          "duplicate" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "StaffDocumentView" : {
        "properties" : {
          "category" : {
            "enum" : [ "DAMAGE_PHOTO", "REPAIR_ESTIMATE", "POLICE_REPORT", "INVOICE", "MEDICAL_REPORT", "OTHER" ],
            "type" : "string"
          },
          "claimId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "contentType" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "declaredContentType" : {
            "type" : "string"
          },
          "docType" : {
            "type" : "string"
          },
          "fileName" : {
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "rejectionReason" : {
            "type" : "string"
          },
          "sha256" : {
            "type" : "string"
          },
          "sizeBytes" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "PENDING_UPLOAD", "UPLOADED", "PROCESSING", "PROCESSED", "FAILED", "REJECTED" ],
            "type" : "string"
          },
          "uploadedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "uploadedBy" : {
            "format" : "int64",
            "type" : "integer"
          },
          "visibleToClaimant" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "StaffUploadResponse" : {
        "properties" : {
          "document" : {
            "$ref" : "#/components/schemas/StaffDocumentView"
          },
          "upload" : {
            "$ref" : "#/components/schemas/UploadInstructions"
          }
        },
        "type" : "object"
      },
...
      "UploadInstructions" : {
        "description" : "PUT the file to uploadUrl with these headers before it expires, then call complete",
        "properties" : {
          "expiresAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "headers" : {
            "additionalProperties" : {
              "type" : "string"
            },
            "type" : "object"
          },
          "method" : {
            "type" : "string"
          },
          "uploadUrl" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "UploadUrlRequest" : {
        "properties" : {
          "category" : {
            "enum" : [ "DAMAGE_PHOTO", "REPAIR_ESTIMATE", "POLICE_REPORT", "INVOICE", "MEDICAL_REPORT", "OTHER" ],
            "type" : "string"
          },
          "contentType" : {
            "description" : "application/pdf, image/jpeg or image/png",
            "example" : "application/pdf",
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "fileName" : {
            "example" : "repair-estimate.pdf",
            "maxLength" : 200,
            "minLength" : 0,
            "type" : "string"
          },
          "sizeBytes" : {
            "description" : "Exact size in bytes; at most 10 MB",
            "example" : 245760,
            "format" : "int64",
            "maximum" : 10485760,
            "minimum" : 1,
            "type" : "integer"
          }
        },
        "required" : [ "category", "contentType", "fileName" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "description" : "Access token from POST /api/v1/auth/login (valid 15 minutes)",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Guidewire-style claims handling: FNOL, exposures, reserves, payments, approvals, SIU and AI document assessment.\n\nAPI version **v1** (contract 1.3.0). Breaking changes get `/api/v2`; additions bump the minor version.",
    "title" : "AI Claims Platform API",
    "version" : "1.3.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/claims/{claimId}/documents" : {
      "get" : {
        "operationId" : "listClaimDocuments",
        "parameters" : [ {
          "in" : "path",
          "name" : "claimId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/StaffDocumentView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "All documents on the claim, including rejected ones",
        "tags" : [ "Documents" ]
      },
      "post" : {
        "operationId" : "startDocumentUpload",
        "parameters" : [ {
          "in" : "path",
          "name" : "claimId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UploadUrlRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffUploadResponse"
                }
              }
            },
            "description" : "Upload URL issued (valid 5 minutes)"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid transition, duplicate or concurrent update"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Well-formed, but breaks a business rule"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Start an upload (staff documents are not shown to the claimant)",
        "tags" : [ "Documents" ]
      }
    },
...
    "/api/v1/documents/{id}/complete" : {
      "post" : {
        "operationId" : "completeDocumentUpload",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/StaffCompletion"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid transition, duplicate or concurrent update"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Well-formed, but breaks a business rule"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Confirm the upload: the file is verified (real type, size, hash)",
        "tags" : [ "Documents" ]
      }
    },
    "/api/v1/documents/{id}/download-url" : {
      "get" : {
        "operationId" : "getDocumentDownloadUrl",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DownloadLinkResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid transition, duplicate or concurrent update"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "A download link valid for 5 minutes",
        "tags" : [ "Documents" ]
      }
    },
...
    "/api/v1/portal/claims/{claimId}/documents" : {
      "get" : {
        "operationId" : "listMyClaimDocuments",
        "parameters" : [ {
          "in" : "path",
          "name" : "claimId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/PortalDocumentView"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Documents on my claim",
        "tags" : [ "Portal: documents" ]
      },
      "post" : {
        "description" : "PUT the file to upload.uploadUrl with upload.headers (the size must match exactly), then call POST /portal/documents/{id}/complete.",
        "operationId" : "startMyDocumentUpload",
        "parameters" : [ {
          "in" : "path",
          "name" : "claimId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        }, {
          "$ref" : "#/components/parameters/CorrelationId"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UploadUrlRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PortalUploadResponse"
                }
              }
            },
            "description" : "Upload URL issued (valid 5 minutes)"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Invalid request: malformed body, bad parameter or failed validation (see `violations`)"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Missing, invalid or expired access token"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "The user's role does not allow this action"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Not found, or not visible to this user"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Conflicts with the current state: invalid transition, duplicate or concurrent update"
          },
          "422" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Well-formed, but breaks a business rule"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                }
              }
            },
            "description" : "Unexpected error (details are logged with the correlation ID, never returned)"
          }
        },
        "summary" : "Start an upload: get a presigned URL",
        "tags" : [ "Portal: documents" ]
      }
    },
...
  },
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
  "servers" : [ {
    "url" : "http://localhost:8081"
  } ],
...
}